requirements.txt
streamlit
tensorflow==2.20.0
keras==3.5.0
keras-vggface==0.6
mtcnn==0.1.1
opencv-python-headless
scikit-learn
numpy
pillow
